[>-<]  merged with PWM_auto

+++ mss_example/SoftConsole/mss_top_sb_MSS_CM3/mss_top_sb_MSS_CM3_app/Helpers/sys_helper/sys_helper.h
@@ -15,13 +15,13 @@ extern "C" {
 
 
 /*
- * time in ms
+ * time in us
  * frequency in MHz
  */
 void delay
 (
-	uint8_t time,
-	uint8_t frequency
+	uint32_t time,
+	uint32_t frequency
 );
 
 #ifdef __cplusplus
